+++ ent/namebhl/interface.go
@@ -3,6 +3,7 @@ package namebhl
 // NameBHL interface provides methods to collect data from BHLindex.
 type NameBHL interface {
 	PageFilesToIDs() error
-	ImportOccurrences() error
 	ImportNames() error
+	ImportOccurrences() error
+	ImportCoLRefs() error
 }