+++ io/namesbhlio/db-namesbhlio.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/dustin/go-humanize"
 	"github.com/lib/pq"
-	"github.com/rs/zerolog/log"
 )
 
 const (
@@ -39,15 +38,15 @@ const (
 
 func (n namesbhlio) saveNames(ch <-chan [][]string) error {
 	total := 0
+	columns := []string{"id", "name", "record_id", "match_type",
+		"edit_distance", "stem_edit_distance", "matched_name", "matched_canonical",
+		"current_name", "current_canonical", "classification",
+		"classification_ranks", "classification_ids", "data_source_id",
+		"data_source_title", "data_sources_number", "curation", "occurences",
+		"odds_log10", "error"}
 
 	for names := range ch {
 		total += len(names)
-		columns := []string{"id", "name", "record_id", "match_type",
-			"edit_distance", "stem_edit_distance", "matched_name", "matched_canonical",
-			"current_name", "current_canonical", "classification",
-			"classification_ranks", "classification_ids", "data_source_id",
-			"data_source_title", "data_sources_number", "curation", "occurences",
-			"odds_log10", "error"}
 		transaction, err := n.db.Begin()
 		if err != nil {
 			return err
@@ -62,32 +61,19 @@ func (n namesbhlio) saveNames(ch <-chan [][]string) error {
 
 		for _, v := range names {
 			eDist, err = strconv.Atoi(v[EditDistanceF])
-			if err != nil {
-				log.Fatal().Err(err).Msg("saveNames:")
-				return err
+			if err == nil {
+				stemDist, err = strconv.Atoi(v[StemEditDistanceF])
 			}
-
-			stemDist, err = strconv.Atoi(v[StemEditDistanceF])
-			if err != nil {
-				log.Fatal().Err(err).Msg("saveNames:")
-				return err
+			if err == nil {
+				dsID, err = strconv.Atoi(v[DataSourceIDF])
 			}
-
-			dsID, err = strconv.Atoi(v[DataSourceIDF])
-			if err != nil {
-				log.Fatal().Err(err).Msg("saveNames:")
-				return err
+			if err == nil {
+				dsNum, err = strconv.Atoi(v[DataSourcesNumberF])
 			}
-
-			dsNum, err = strconv.Atoi(v[DataSourcesNumberF])
-			if err != nil {
-				log.Fatal().Err(err).Msg("saveNames:")
-				return err
+			if err == nil {
+				occurs, err = strconv.Atoi(v[OccurrencesNumberF])
 			}
-
-			occurs, err = strconv.Atoi(v[OccurrencesNumberF])
 			if err != nil {
-				log.Fatal().Err(err).Msg("saveNames:")
 				return err
 			}
 
@@ -102,7 +88,6 @@ func (n namesbhlio) saveNames(ch <-chan [][]string) error {
 				v[ClassificationF], v[ClassificationRanksF], v[ClassificationIDsF],
 				dsID, v[DataSourceF], dsNum, true, occurs, odds, v[ErrorF])
 			if err != nil {
-				log.Fatal().Err(err).Msg("saveNames:")
 				return err
 			}
 		}