+++ io/namesbhlio/db-occurs.go
@@ -11,8 +11,6 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-const OccurBatchSize = 50_000
-
 func (n namesbhlio) saveOcurrences(chIn <-chan []db.NameOccurrence) error {
 	columns := []string{"page_id", "name_string_id", "offset_start",
 		"offset_end", "odds_log10", "annot_nomen"}
@@ -20,12 +18,11 @@ func (n namesbhlio) saveOcurrences(chIn <-chan []db.NameOccurrence) error {
 	for ocs := range chIn {
 		transaction, err := n.db.Begin()
 		if err != nil {
-			err = fmt.Errorf("saveOccurrences: %w", err)
 			return err
 		}
 		stmt, err := transaction.Prepare(pq.CopyIn("name_occurrences", columns...))
 		if err != nil {
-			return fmt.Errorf("saveOccurrences: %w", err)
+			return err
 		}
 
 		for i := range ocs {
@@ -33,23 +30,23 @@ func (n namesbhlio) saveOcurrences(chIn <-chan []db.NameOccurrence) error {
 				ocs[i].OffsetStart, ocs[i].OffsetEnd, ocs[i].OddsLog10,
 				ocs[i].AnnotNomen)
 			if err != nil {
-				return fmt.Errorf("saveOccurrences: %w", err)
+				return err
 			}
 		}
 
 		count += len(ocs)
-		missing += OccurBatchSize - len(ocs)
+		missing += occurBatchSize - len(ocs)
 		fmt.Fprintf(os.Stderr, "\r%s", strings.Repeat(" ", 47))
 		fmt.Fprintf(os.Stderr, "\rImported %s occurrences.", humanize.Comma(int64(count)))
 
 		err = stmt.Close()
 		if err != nil {
-			return fmt.Errorf("saveOccurrences: %w", err)
+			return err
 		}
 
 		err = transaction.Commit()
 		if err != nil {
-			return fmt.Errorf("saveOccurrences: %w", err)
+			return err
 		}
 	}
 	fmt.Fprintln(os.Stderr)